refactor(FeedbackRules): Create FeedbackRuleExpression class to handle postfix conversion
